Start Freenet from Supervisor
The daemon is run in the ``CENOBridge`` since it already contains the
``freenet.ini`` file, but it may not be the most adequate place to run it
since it fills the directory with work files.

It also complains about the missing wrapper, but I think it's ok since we're
using Supervisor for that.  It may also have problems with missing plugins.